Develop (#6)
* added local variables for violiplot function

* moved to R 4.3, updated packages

* added iqr of beta to stored beta thresholds

* added 1st quantile and 3rd quantile to stored beta thresholds

* added test for new data threshold for beta range

* improved global status management
testing using multicore
added deltar and deltaqr calculation

* removed beta saving

* documented functions
removed Rfast unused library

* added beta management

* removed reuse of existing annotated bed file

* added management of BETA values

* improved session info

* refactored names of anomaly, population, figure, genomic area identification
Renamed test to be executed from basic to highest level

* renamed POPULATION as Sample_Group, refactoring

* changing to manage the new method to manage annotations

* renamed analize population, corrected problem artifact with setup for testing

* refactoring

* moved tempo folder setup into test setup

change annotate_bed to run all the file to annotate using the config setup of environment
changed create_excel_pivot to run all the file to annotate using the config setup of environment
changed create_heatmap to run all the file to annotate using the config setup of environment

introduced progeress bar

* debugged create excel_pivot
improved management of BETA keys

* added reshape with renv and in DESCRIPTION

* debigged annotate_bed moving colname definition after control existence of final_bed

removed unused variable in create_heatmap

removed test of more than one group in annotate_bed test file

* renamed ANNOTATED to Annotated, debigged update session checking existence of folder and not file.
Improved test of annotated and create excel pivot
Completed excel pivot.

* refactored tests to have a temp folder exclusively per each test

* corrected cmd check warning, removed parallel execution of tests

* documented annotation data

* documented pp_tot

* removed debian as testing and added ubuntu

---------

+++ R/create_multiple_bed.R
@@ -1,58 +1,63 @@
 #' @importFrom doRNG %dorng%
 create_multiple_bed <- function(sample_sheet){
 
-  ssEnv <- .pkgglobalenv$ssEnv
-
-  message("INFO: ", Sys.time(), " Started multiple annotated file creation!")
-
+  ssEnv <- get_session_info()
+  message("INFO: ", Sys.time(), " Started multiple file creation!")
   #create multiple file bed
   i <- 0
-  localKeys <- expand.grid("POPULATION"=unique(sample_sheet$Sample_Group),
-                           "FIGURE"=ssEnv$keys_figures_default[,1] ,
-                           "ANOMALY"= c("MUTATIONS","LESIONS") ,"EXT"="bed")
-  localKeys <- rbind(localKeys, expand.grid("POPULATION"=unique(sample_sheet$Sample_Group),
-                                            "FIGURE"=ssEnv$keys_figures_default[,1],
-                                            "ANOMALY"="DELTAS" ,"EXT"="bedgraph"))
+  Sample_Group <- as.data.frame(unique(sample_sheet$Sample_Group))
+  colnames(Sample_Group) <- "SAMPLE_GROUP"
+  localKeys <- reshape::expand.grid.df(ssEnv$keys_markers_figures,Sample_Group)
+  # localKeys <- subset(localKeys, MARKER!="BETA")
 
-  # to_export <- c("localKeys", "dir_check_and_create", "ssEnv", "file_path_build", "sample_sheet")
-  future::plan( future::sequential)
-  # foreach::foreach(i = 1:nrow(localKeys), .export = to_export) %dorng%
-  for(i in 1:nrow(localKeys))
+  if(ssEnv$showprogress)
+    progress_bar <- progressr::progressor(along = 1:nrow(localKeys))
+  else
+    progress_bar <- ""
+
+  to_export <- c("localKeys", "dir_check_and_create", "ssEnv", "file_path_build", "sample_sheet","progress_bar",
+    "progression_index", "progression", "progressor_uuid", "owner_session_uuid", "trace")
+  # future::plan( future::sequential)
+  foreach::foreach(i = 1:nrow(localKeys), .export = to_export) %dorng%
+  # for(i in 1:nrow(localKeys))
   {
+    # i <- 1
     key <- localKeys[i,]
-    tempresult_folderData <-dir_check_and_create(ssEnv$result_folderData,c(as.character(key$POPULATION) ,paste(as.character(key$ANOMALY),"_",as.character(key$FIGURE),sep="")))
+    marker <- as.character(key$MARKER)
+    figure <- as.character(key$FIGURE)
+    tempresult_folderData <-dir_check_and_create(ssEnv$result_folderData,c(as.character(key$SAMPLE_GROUP) ,paste(as.character(marker),"_",as.character(figure),sep="")))
     temp_file <- tempdir()
     temp_file <- paste(temp_file, stringi::stri_rand_strings(1, 12, pattern = "[A-Za-z0-9]"),sep="")
-    fileToWrite <- file_path_build(tempresult_folderData, c("MULTIPLE", as.character(key$ANOMALY), as.character(key$FIGURE)), "fst")
-    j <- 0
-    for ( j in 1:nrow(sample_sheet))
+    fileToWrite <- file_path_build(tempresult_folderData, c("MULTIPLE", as.character(marker), as.character(figure)), "fst")
+    if(!file.exists(fileToWrite))
     {
-      sample <- sample_sheet[j,]
-      fileToRead <- file_path_build(tempresult_folderData, c(sample$Sample_ID, as.character(key$ANOMALY), as.character(key$FIGURE)), key$EXT)
-      if(file.exists(fileToRead))
+      j <- 0
+      for ( j in 1:nrow(sample_sheet))
       {
-        localtemp <- utils::read.csv2(fileToRead, sep="\t", header = FALSE)
-        localtemp$Sample_ID <- sample$Sample_ID
-        if(!plyr::empty(localtemp))
+        # j <- 1
+        sample <- sample_sheet[j,]
+        fileToRead <- file_path_build(tempresult_folderData, c(sample$Sample_ID, as.character(marker), as.character(figure)), key$EXT)
+        if(file.exists(fileToRead))
         {
-          utils::write.table(localtemp, temp_file, sep = "\t", row.names = FALSE, col.names = FALSE, quote = FALSE, append = TRUE)
+          localtemp <- utils::read.csv2(fileToRead, sep="\t", header = FALSE)
+          localtemp$Sample_ID <- sample$Sample_ID
+          if(!plyr::empty(localtemp))
+          {
+            utils::write.table(localtemp, temp_file, sep = "\t", row.names = FALSE, col.names = FALSE, quote = FALSE, append = TRUE)
+          }
+          rm(localtemp)
         }
-        rm(localtemp)
       }
-    }
 
-    if(file.exists(temp_file))
-    {
-        # if(ncol(temp_file)==4)
-        #   colnames(temp_file)<- c("CHR","START","END","SAMPLEID")
-        # else
-        #   colnames(temp_file)<- c("CHR","START","END","VALUE","SAMPLEID")
-
+      if(file.exists(temp_file))
+      {
         fst::write.fst(x = utils::read.table(temp_file, sep = "\t"),path = fileToWrite)
         file.remove(temp_file)
         message("INFO: ", Sys.time(), " Created multiple annotated file!", fileToWrite)
+      }
     }
-    gc()
+    if(ssEnv$showprogress)
+      progress_bar(sprintf("Creating multiple file."))
   }
-  future::plan( ssEnv$parallel_strategy)
+  # future::plan(ssEnv$parallel_strategy)
 }
